requirements.txt
numpy
scipy
scikit-learn
tqdm
pandas
PyYAML
openai==1.93.0
transformers>=4.51.1
accelerate>=0.26.0
google-genai
# Optional: For Gemini embeddings
# To remove the old package: pip uninstall google-generativeai
# To install the new package: pip install google-genai
pytest
